If weights for some buckets are equal, result depends on the
initial positions of buckets in slice.

Props to @ZhangTao1596.

Signed-off-by: Evgenii Stratonikov <evgeniy@nspcc.ru>