@fyrchik 2021-12-27
Sorting by HRW value (instead of index) makes `GetContainerNodes` a bit faster.
```
name             old time/op    new time/op    delta
PolicyHRWType-8    53.7µs ± 4%    51.9µs ± 4%  -3.49%  (p=0.000 n=19+19)

name             old alloc/op   new alloc/op   delta
PolicyHRWType-8    21.6kB ± 0%    20.6kB ± 0%  -4.70%  (p=0.000 n=19+20)

name             old allocs/op  new allocs/op  delta
PolicyHRWType-8       421 ± 0%       407 ± 0%  -3.33%  (p=0.000 n=20+20)
```