@alexvanin 2021-12-27
Sorting by value is the way to go. Should be there from the start.