@fyrchik 2021-12-28
I figured out, there was a mistake in benchmarks. HRW library silently returns when elements of slice provided to `SortSliceByWeightValue` do not implement `Hasher` interface. After I had added `Nodes.Hash` method, I got following results (`netmapSize = 1000`):
```
BenchmarkHRWSort/sort_by_index,_no_weight
BenchmarkHRWSort/sort_by_index,_no_weight-8         	    8715	    133716 ns/op
BenchmarkHRWSort/sort_by_value,_no_weight
BenchmarkHRWSort/sort_by_value,_no_weight-8         	    6654	    186734 ns/op
BenchmarkHRWSort/only_sort_by_index
BenchmarkHRWSort/only_sort_by_index-8               	    5784	    211262 ns/op
BenchmarkHRWSort/sort_by_value
BenchmarkHRWSort/sort_by_value-8                    	    4398	    262967 ns/op
BenchmarkHRWSort/sort_by_ID,_then_by_index_(deterministic)
BenchmarkHRWSort/sort_by_ID,_then_by_index_(deterministic)-8         	    4540	    262614 ns/op
```
So it turns out sorting by value is a bit slower than sorting by index. But we really need to compare implementations with _deterministic_ results (last 2 lines). And here the difference is not so significant. If we set `netmapSize = 10_000` in benchmarks, `sort_by_value` starts to be faster.